+++ miamCore/src/commonMain/kotlin/com/miam/kmmMiamCore/component/itemSelector/ItemSelectorContract.kt
@@ -1,5 +1,6 @@
 package com.miam.kmmMiamCore.component.itemSelector
 
+import com.miam.kmmMiamCore.base.mvi.BasicUiState
 import com.miam.kmmMiamCore.base.mvi.UiEffect
 import com.miam.kmmMiamCore.base.mvi.UiEvent
 import com.miam.kmmMiamCore.base.mvi.UiState
@@ -8,26 +9,21 @@ import com.miam.kmmMiamCore.miam_core.model.BasketPreviewLine
 
 interface ItemSelectorContract {
 
-    sealed class Event : UiEvent {
-        object ReturnToBasketPreview : ItemSelectorContract.Event()
-        data class SetReturnToBasketPreview(val returnToPreview: () -> Unit) :
-            ItemSelectorContract.Event()
-
-        data class SetItemList(val items: List<BasketPreviewLine>) : ItemSelectorContract.Event()
-        data class SetSelectedItem(val item: BasketPreviewLine) : ItemSelectorContract.Event()
-        data class SetReplaceItemInPreview(val replace: (be: BasketEntry) -> Unit) :
-            ItemSelectorContract.Event()
-
-        data class SelectNewItem(val index: Int) : ItemSelectorContract.Event()
+    sealed class Event: UiEvent {
+        object ReturnToBasketPreview: Event()
+        data class SetReturnToBasketPreview(val returnToPreview: () -> Unit): Event()
+        data class SetItemList(val items: List<BasketPreviewLine>): Event()
+        data class SetReplaceItemInPreview(val replace: (be: BasketEntry) -> Unit): Event()
+        data class SelectNewItem(val index: Int): Event()
     }
 
     data class State(
-        val selectedItem: BasketPreviewLine? = null,
-        val itemList: List<BasketPreviewLine>? = mutableListOf(),
+        val selectedItem: BasicUiState<BasketPreviewLine>,
+        val items: List<BasketPreviewLine> = mutableListOf(),
         val replaceItemInPreview: (be: BasketEntry) -> Unit,
         val returnToPreview: () -> Unit
-    ) : UiState
+    ): UiState
 
-    sealed class Effect : UiEffect
+    sealed class Effect: UiEffect
 }