+++ changelog.md
@@ -1,5 +1,7 @@
 # 3.8.1
 - [FIX][AND] Add a loader on guest counter in basket preview
+- [FIX][IOS] Add a loader on guest counter in basket preview and on items row
+- [FIX][IOS] Counter in recipe detail can't go under 1 add a isReloading parameter to template
 
 # 3.8.0
 - [FEA][COR] Expose analytic events and closure