+++ changelog.md
@@ -1,5 +1,6 @@
-# 3.9.0
+# 3.8.1
 - [FIX][AND] Prefrence search modal now handle back press correctly
+- [FIX][AND] Fix crash on itemSelector openning
 # 3.8.0
 - [FEA][COR] Expose analytic events and closure
 - [FEA][IOS] Use native navigation for recipe details, basket preview, basket tags list and item selector